# Refactor the merkle tree to refactor the last siblings in memory