Closes #283
### What does this PR do?

<!-- Describe your changes here- ideally you can get that description straight from
your descriptive commit message(s)! -->

### Reviewers

Main reviewers:

<!-- Main reviewers should do a full review. There should be 2 main reviewers, unless there is a good reason for not to do it -->

- @ARR552 
- @arnaubennassar 


### Benchmarking
- old one

```
cpu: Intel(R) Core(TM) i5-9400 CPU @ 2.90GHz
BenchmarkApiSmallTest/test-sync-postgres-500-6                 1        18450188585 ns/op
BenchmarkApiSmallTest/test-api-postgres-500/get_bridges_endpoint-6              1000000000               0.01291 ns/op         0 B/op          0 allocs/op
BenchmarkApiSmallTest/test-api-postgres-500/get_merkle_proof_endpoint-6         1000000000               0.05116 ns/op         0 B/op          0 allocs/op
```

- new one

```
cpu: Intel(R) Core(TM) i5-9400 CPU @ 2.90GHz
BenchmarkApiSmallTest/test-sync-postgres-500-6                 1        2450893724 ns/op
BenchmarkApiSmallTest/test-api-postgres-500/get_bridges_endpoint-6              1000000000               0.01451 ns/op         0 B/op          0 allocs/op
BenchmarkApiSmallTest/test-api-postgres-500/get_merkle_proof_endpoint-6         1000000000               0.04975 ns/op         0 B/op          0 allocs/op
```